RectGrid_test: Parametrize data written to test file
In future parallel test will write a file in serial mode,
will read it in parallel mode and will write it again in
parallel mode. In both write the date should be set
to the same value to make comparison of files easier. Therefore
the date should be parametrized by a constant.